fixed some errors
