+++ app.vue
@@ -1,3 +1,7 @@
+<script setup lang="ts">
+import ResourceCreate from './components/resourceCreate.vue';
+</script>
+
 <template>
   <div class="font-bold">
     <NuxtRouteAnnouncer />
@@ -6,5 +10,9 @@
   <NuxtPage />
   <ResourceEdit/>
   <ResourceCard/>
+  <ResourceCreate/>
+
+  
+  
 </template>
