+++ components/resourceEdit.vue
@@ -1,90 +1,68 @@
+<script setup lang="ts">
+import { useFetch } from 'nuxt/app';
+import { ref } from 'vue';
+
+const editMode = ref(false);
+const resourceId = 3; // Set this dynamically if needed
+
+// Fetch the resource data
+const { data: resource, error, refresh } = await useFetch(`/api/resource/get/${resourceId}`);
+
+// Function to save the updated resource
+const saveChanges = async () => {
+  if (resource.value) {
+    try {
+      const response = await fetch(`http://localhost:3000/api/resource/put/${resourceId}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          name: resource.value.name,
+          description: resource.value.description,
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to update resource');
+      }
+
+      // Option 1: Refresh the resource to get the updated data
+      await refresh();
+
+      // Option 2: Manually update the resource (if you don't want to make another GET request)
+      // resource.value.name = updatedData.name;
+      // resource.value.description = updatedData.description;
+
+      // Disable edit mode after saving
+      editMode.value = false;
+      console.log('Resource updated successfully');
+    } catch (err) {
+      console.error('Error saving changes:', err);
+    }
+  }
+};
+</script>
+
 <template>
-  <div class="resource-card">
-    <h2>Create Resource</h2>
-    <form @submit.prevent="submitForm">
-      <div>
-        <label for="name">Name:</label>
-        <input v-model="resource.name" type="text" id="name" required />
-      </div>
-      <div>
-        <label for="description">Description:</label>
-        <textarea v-model="resource.description" id="description" required></textarea>
-      </div>
-      <div>
-        <label for="group">Group:</label>
-        <input v-model="resource.group" type="text" id="group" required />
-      </div>
-      <!-- Add other fields as needed -->
-      <button type="submit">Create Resource</button>
-    </form>
+  <div v-if="error">
+    <p>Error: {{ error.message }}</p>
+  </div>
+  <div v-else-if="resource && 'error' in resource">
+    <p>Error: {{ resource.error }}</p>
+  </div>
+  <div v-else-if="!resource">
+    <p>Loading...</p>
+  </div>
+  <div v-else>
+    <button @click="editMode ? saveChanges() : editMode = true">
+      {{ editMode ? 'Save' : 'Edit' }}
+    </button>
 
-    <div v-if="successMessage">
-      <p>{{ successMessage }}</p>
-    </div>
-
-    <div v-if="errorMessage">
-      <p>{{ errorMessage }}</p>
+    <!-- Editable fields for name and description in edit mode -->
+    <div v-if="editMode" class="mt-2">
+      <input v-model="resource.name" type="text" placeholder="Edit Name" />
+      <input v-model="resource.description" type="text" placeholder="Edit Description" />
     </div>
   </div>
 </template>
-
-<script lang="ts">
-import { defineComponent, ref } from 'vue';
-
-// Define TypeScript interfaces for the resource form
-interface Resource {
-  name: string;
-  description: string;
-  group: string;
-  // Optionally, other fields can be added here
-}
-
-export default defineComponent({
-  setup() {
-    // State variables
-    const resource = ref<Resource>({
-      name: '',
-      description: '',
-      group: '',
-    });
-
-    const successMessage = ref<string>('');
-    const errorMessage = ref<string>('');
-
-    // Function to handle form submission
-    const submitForm = async () => {
-      try {
-        const response = await fetch('/api/resource/post/one', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(resource.value),
-        });
-
-        const data = await response.json();
-
-        if (response.ok && data.success) {
-          successMessage.value = `Resource created successfully! ID: ${data.resource.id}`;
-        } else {
-          errorMessage.value = data.error || 'Failed to create resource';
-        }
-      } catch (error) {
-        errorMessage.value = 'An error occurred while creating the resource';
-        console.error(error);
-      }
-    };
-
-    return {
-      resource,
-      successMessage,
-      errorMessage,
-      submitForm,
-    };
-  },
-});
-</script>
-
-<style scoped>
-/* Add styles for the form */
-</style>
